put query error methods into their own class added responses to exception controller fixed errors in README
