+++ minianalytics-api/src/main/java/amcmurray/bw/controllers/ExceptionController.java
@@ -1,21 +1,25 @@
 package amcmurray.bw.controllers;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import amcmurray.bw.twitterdomainobjects.Query;
+import amcmurray.bw.exceptions.QueryExceptions;
 
 @ControllerAdvice
 @RestController
 public class ExceptionController {
 
-    @ExceptionHandler(Query.QueryNotFoundException.class)
-    public String queryNotFoundHandler(Query.QueryNotFoundException e) {
+    @ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "Query not found.")  // 404 error
+    @ExceptionHandler(QueryExceptions.QueryNotFoundException.class)
+    public String queryNotFoundHandler(QueryExceptions.QueryNotFoundException e) {
         return "Query not found with ID: " + e.id;
     }
 
-    @ExceptionHandler(Query.QuerySearchNullException.class)
+    @ResponseStatus(value = HttpStatus.BAD_REQUEST, reason = "Query text must not be empty!")  // 400 error
+    @ExceptionHandler(QueryExceptions.QuerySearchNullException.class)
     public String querySearchNullException() {
         return "Query text must not be empty!";
     }
