+++ minianalytics-api/src/main/java/amcmurray/bw/exceptions/QueryExceptions.java
@@ -1,0 +1,17 @@
+package amcmurray.bw.exceptions;
+
+public class QueryExceptions extends Exception {
+
+    public static class QueryNotFoundException extends RuntimeException {
+        public int id;
+
+        public QueryNotFoundException(int id) {
+            this.id = id;
+        }
+    }
+
+    public static class QuerySearchNullException extends RuntimeException {
+    }
+
+
+}
